chore: Home의 UiState, Event, SideEffect model 패키지로 이동
