+++ android/feature/home/src/main/kotlin/com/andlife/home/model/HomeSideEffect.kt
@@ -2,4 +2,8 @@ package com.andlife.home.model
 
 import com.andlife.ui.base.BaseSideEffect
 
-sealed interface HomeSideEffect : BaseSideEffect
+sealed interface HomeSideEffect : BaseSideEffect {
+    data class ShowMessage(
+        val message: String,
+    ) : HomeSideEffect
+}
