+++ android/feature/home/src/main/kotlin/com/andlife/home/model/HomeUiState.kt
@@ -1,7 +1,9 @@
 package com.andlife.home.model
 
+import com.andlife.model.guestbook.GuestBookUiModel
 import com.andlife.ui.base.BaseUiState
 
 data class HomeUiState(
     val isLoading: Boolean = false,
+    val guestBooks: List<GuestBookUiModel> = emptyList(),
 ) : BaseUiState
